Big refactor
Moved segmentation into own package
Moved test data
Moved weights to Zenodo
