+++ renal_segmentor.py
@@ -8,135 +8,14 @@ Created on Fri Jul 12 11:25:34 2019
 # Import Packages
 
 import nibabel as nib
-import numpy as np
 import os
 import pandas as pd
 import sys
 import tensorflow as tf
 
 from gooey import Gooey, GooeyParser
-from nibabel.processing import conform
-from skimage.measure import label, regionprops
-from skimage.transform import resize
-from tensorflow.keras import backend as K
-from tensorflow.keras.models import load_model
-
-# Define Classes
-
-
-class RawData:
-    def __init__(self, path):
-        self.path = path
-        self.__split_path__()
-        self.img = nib.Nifti1Image
-        self.data = np.array
-        self.mask = np.array
-        self.affine = np.array
-        self.shape = tuple
-        self.zoom = tuple
-        self.tkv = np.nan
-        self.lkv = np.nan
-        self.rkv = np.nan
-
-    def __split_path__(self):
-        self.directory = os.path.dirname(self.path)
-        self.base = os.path.splitext(os.path.basename(self.path))[0]
-        self.extension = os.path.splitext(os.path.basename(self.path))[1]
-        if self.extension == '.gz' and self.base[-4:] == '.nii':
-            self.extension = '.nii.gz'
-            self.base = self.base[:-4]
-
-    def load(self):
-        if self.extension == '.PAR':
-            self.img = nib.load(self.path, scaling='fp')
-        else:
-            self.img = nib.load(self.path)
-        self.data = self.img.get_fdata()
-        self.affine = self.img.affine
-        self.shape = self.img.shape
-        self.zoom = self.img.header.get_zooms()
-        self.orientation = nib.orientations.aff2axcodes(self.affine)
-
-    def get_mask(self, post_process=True,
-                 weights_path='./models/renal_segmentor.model'):
-        img = conform(self.img, out_shape=(240, 240, self.shape[-1]),
-                      voxel_size=(1.458, 1.458, self.zoom[-1] * 0.998),
-                      orientation='LIP')
-        data = img.get_fdata()
-        data = np.flip(data, 1)
-        data = np.swapaxes(data, 0, 2)
-        data = np.swapaxes(data, 1, 2)
-        data = self._rescale(data)
-        data = resize(data, (data.shape[0], 256, 256))
-        data = data.reshape((data.shape[0], data.shape[1], data.shape[2], 1))
-        model = load_model(resource_path(weights_path),
-                           custom_objects={'dice_coef_loss':
-                                           self._dice_coef_loss,
-                                           'dice_coef': self._dice_coef})
-        batch_size = 2 ** 3
-        mask = model.predict(data, batch_size=batch_size)
-        mask = np.squeeze(mask)
-        mask = np.swapaxes(mask, 0, 2)
-        mask = np.swapaxes(mask, 0, 1)
-        mask = np.flip(mask, 1)
-        mask = resize(mask, (240, 240, self.shape[-1]))
-        if post_process:
-            cleaned_mask = self._cleanup(mask > 0.05)
-            mask[cleaned_mask < 0.5] = 0.0
-        mask_img = nib.Nifti1Image(mask, img.affine)
-        mask_img = conform(mask_img, out_shape=self.shape,
-                           voxel_size=self.zoom, orientation=self.orientation)
-        self.mask = self._rescale(mask_img.get_fdata(), 0, 1)
-        self.tkv = (np.sum(self.mask > 0.5) *
-                    np.prod(self.zoom))/1000
-        self.lkv = (np.sum(self.mask[120:] > 0.5) *
-                    np.prod(self.zoom))/1000
-        self.rkv = (np.sum(self.mask[:120] > 0.5) *
-                    np.prod(self.zoom)) / 1000
-        return self.mask
-
-    @staticmethod
-    def _rescale(data, black=None, white=None):
-        if black is None:
-            black = np.mean(data) - 0.5 * np.std(data)
-            if black < data.min():
-                black = data.min()
-        if white is None:
-            white = np.mean(data) + 4 * np.std(data)
-            if white > data.max():
-                white = data.max()
-        data = np.clip(data, black, white) - black
-        data = data / (white - black)
-        return data
-
-    @staticmethod
-    def _cleanup(mask):
-        clean_mask = np.zeros(mask.shape, dtype=np.uint8)
-        label_mask = label(mask > 0.5, connectivity=1)
-        props = regionprops(label_mask)
-        areas = [region.area for region in props]
-
-        # This means there have to be two kidneys in the scan...
-        kidney_labels = np.argpartition(areas, -2)[-2:]
-
-        clean_mask[label_mask == props[kidney_labels[0]].label] = 1
-        clean_mask[label_mask == props[kidney_labels[1]].label] = 1
-
-        return clean_mask
-
-    @staticmethod
-    def _dice_coef(y_true, y_pred):
-        smooth = 1.0
-        y_true_f = K.flatten(y_true)
-        y_pred_f = K.flatten(y_pred)
-        intersection = K.sum(y_true_f * y_pred_f)
-        return (2. * intersection + smooth) / (
-                    K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
-
-    def _dice_coef_loss(self, y_true, y_pred):
-        loss = 1 - self._dice_coef(y_true, y_pred)
-        return loss
-
+from segment import Tkv
+from segment.data import fetch
 
 # Define Functions
 
@@ -265,38 +144,39 @@ def main():
 
     parser = get_parser()
     args = parser.parse_args()
-
+    weights_path = resource_path(fetch.Weights().path)
     # Import data
     volumes = pd.DataFrame(index=args.input, columns=['tkv', 'lkv', 'rkv'])
     n = 0
     for path, row in volumes.iterrows():
         n += 1
-        raw_data = RawData(path)
-        raw_data.load()
+        segmentation = Tkv(path)
+        segmentation.load()
 
-        mask = raw_data.get_mask(post_process=args.post_process)
+        mask = segmentation.get_mask(post_process=args.post_process,
+                                     weights_path=weights_path)
 
-        volumes.loc[path, 'tkv'] = raw_data.tkv
-        volumes.loc[path, 'lkv'] = raw_data.lkv
-        volumes.loc[path, 'rkv'] = raw_data.rkv
+        volumes.loc[path, 'tkv'] = segmentation.tkv
+        volumes.loc[path, 'lkv'] = segmentation.lkv
+        volumes.loc[path, 'rkv'] = segmentation.rkv
 
         if args.binary:
             mask = (mask > 0.5) * 1
 
         # Output mask
         if not args.output:
-            out_dir = raw_data.directory
+            out_dir = segmentation.directory
         else:
             out_dir = args.output
 
-        mask_fname = os.path.join(out_dir, raw_data.base + '_mask.nii.gz')
+        mask_fname = os.path.join(out_dir, segmentation.base + '_mask.nii.gz')
 
-        mask_img = nib.Nifti1Image(mask, raw_data.affine)
+        mask_img = nib.Nifti1Image(mask, segmentation.affine)
         nib.save(mask_img, mask_fname)
 
         if args.raw:
-            nib.save(raw_data.img, os.path.join(out_dir, raw_data.base +
-                                                '.nii.gz'))
+            nib.save(segmentation.img, os.path.join(out_dir, segmentation.base
+                                                    + '.nii.gz'))
 
         print(f'Processed {n} of {len(volumes)} files')
