+++ requirements.txt
@@ -6,3 +6,4 @@ pyinstaller==4.5.1
 pytest==6.2.5
 scikit-image==0.18.3
 tensorflow==2.6.0
+wget==3.2
