+++ segment/data/test_data/test_sub_02.nii.gz
