+++ segment/tests/utils.py
@@ -1,0 +1,18 @@
+import numpy as np
+import numpy.testing as npt
+
+
+def image_stats(data):
+    mean = np.nanmean(data)
+    std = np.nanstd(data)
+    max = np.nanmax(data)
+    min = np.nanmin(data)
+    dim = data.shape[-1]
+    row_sum = np.sum(data, axis=0).reshape(-1)[0]
+    return [mean, std, max, min, dim, row_sum]
+
+
+def same_image(test, gold_image_stats):
+    if type(test) != list and len(test) != 6:
+        test = image_stats(test)
+    npt.assert_allclose(test, gold_image_stats, rtol=1e-3, atol=1e-9)
